Removed several Sonarlint warnings

+++ src/main/java/com/github/robtimus/io/stream/BinaryPipe.java
@@ -174,7 +174,7 @@ public final class BinaryPipe {
             }
             throwWriteError();
             if (start < end) {
-                long result = Math.min(n, end - start);
+                long result = Math.min(n, (long) end - start);
                 start += result;
                 checkEmpty();
                 return result;
@@ -380,6 +380,7 @@ public final class BinaryPipe {
 
     private void await(Condition condition) throws IOException {
         try {
+            // if this call waits shorter than AWAIT_TIME that's fine, the check will simply be performed earlier
             condition.awaitNanos(AWAIT_TIME);
         } catch (InterruptedException e) {
             // Restore the interrupted status
